Allow creating fonts with `Arc<dyn AsRef[u8] + Send + Sync>` (#57)
This has advantages over a simple `Arc<Vec<u8>>` because it can avoid
data copies. A common case that this happens is when Servo creates
fonts from shared memory. In addition, this maintains full compatability
with `Arc<Vec<u8>>` because it fulfills the trait.

Finally, this bumps the version so that the changes can be deployed.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>